Reuse presenters from spree core for filters
