+++ api/app/presenters/spree/api/products/filters_presenter.rb
@@ -8,11 +8,12 @@ module Spree
 
         def to_h
           option_values = Spree::OptionValues::FindAvailable.new(products_scope: products_for_filters).execute
+          option_values_presenters = Spree::Filters::OptionsPresenter.new(option_values_scope: option_values).to_a
           product_properties = Spree::ProductProperties::FindAvailable.new(products_scope: products_for_filters).execute
-
+          product_properties_presenters = Spree::Filters::PropertiesPresenter.new(product_properties_scope: product_properties).to_a
           {
-            option_types: Spree::Products::OptionTypeFiltersPresenter.new(option_values).to_a,
-            product_properties: Spree::Products::PropertyFiltersPresenter.new(product_properties).to_a
+            option_types: option_values_presenters.map(&:to_h),
+            product_properties: product_properties_presenters.map(&:to_h)
           }
         end
